RESOLVED
Bug Description: The interpolation routine was catered towards 256x256 images in two lines of code (17 & 18).
When we calculated the "nw" of each point, we wanted to be capable of handling a "nw" that is on the edge, meaning that "se" for example would be outside of image's boundary.
Apparently the check we were doing was good only for 256x256 images, and in our examples we only tried 256x256 images so it was fine.
Now we changed this check to take care of all sorts of heights and widths of an image.
P.S: we could've just removed lines 17 & 18 and it would've worked well because we don't really need to check if the "nw" of a point was on the edge, for this is a HIGHLY unlikely situation (it's very hard for our eyes and fingers to chose pixels that result in this situation).